# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Blue Media social platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: Get home feed posts
      parameters:
        - in: query
          name: userId
          schema:
            type: integer
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /posts/{id}/reactions:
    post:
      operationId: reactToPost
      tags: [posts]
      summary: React to a post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: Reaction set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: removePostReaction
      tags: [posts]
      summary: Remove reaction from post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reaction removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}/comments:
    get:
      operationId: listPostComments
      tags: [posts]
      summary: List comments on a post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: addPostComment
      tags: [posts]
      summary: Add a comment to a post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /posts/{id}/comments/{commentId}:
    delete:
      operationId: deletePostComment
      tags: [posts]
      summary: Delete a comment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: commentId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        profilePicture:
          type: ["string", "null"]
        coverPicture:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        privacy:
          type: string
          enum: [public, friends_of_friends, private]
        isAdmin:
          type: boolean
        isBlueAI:
          type: boolean
        blueBadge:
          type: boolean
        blueBadgeClaimedAt:
          type: ["string", "null"]
          format: date-time
        restricted:
          type: boolean
        banned:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        privacy:
          type: string
          enum: [public, friends_of_friends, private]
...
    Post:
      type: object
      required: [id, userId, content, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        bgColor:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        liveStreamId:
          type: ["string", "null"]
        visibility:
          type: string
          enum: [public, friends_of_friends, private]
        author:
          $ref: "#/components/schemas/User"
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/PostReaction"
        commentCount:
          type: integer
        myReaction:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PostInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        bgColor:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        liveStreamId:
          type: ["string", "null"]
        feeling:
          type: ["string", "null"]
        activity:
          type: ["string", "null"]
        locationTag:
          type: ["string", "null"]
        visibility:
          type: string
          enum: [public, friends_of_friends, private]
    PostReaction:
      type: object
      required: [type, count]
      properties:
        type:
          type: string
        count:
          type: integer
    ReactionInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: ["heart", "cry", "laugh", "angry"]
    Comment:
      type: object
      required: [id, postId, userId, content, createdAt]
      properties:
        id:
          type: integer
        postId:
          type: integer
        userId:
          type: integer
        content:
          type: string
        author:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
...