# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Blue Media social platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PostReaction } from './postReaction';
import type { PostVisibility } from './postVisibility';
import type { User } from './user';

export interface Post {
  id: number;
  userId: number;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  bgColor?: string | null;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  liveStreamId?: string | null;
  visibility?: PostVisibility;
  author?: User;
  reactions?: PostReaction[];
  commentCount?: number;
  /** @nullable */
  myReaction?: string | null;
  createdAt: Date;
}