# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Blue Media social platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PostInputVisibility } from './postInputVisibility';

export interface PostInput {
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  bgColor?: string | null;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  liveStreamId?: string | null;
  /** @nullable */
  feeling?: string | null;
  /** @nullable */
  activity?: string | null;
  /** @nullable */
  locationTag?: string | null;
  visibility?: PostInputVisibility;
}